chore: move lockfile code to internal module
In a future commit, I'll need the lockfile code to access the config
file. This solves an import cycle.
