+++ cmd/check.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"github.com/cupcakearmy/autorestic/internal"
 	"github.com/cupcakearmy/autorestic/internal/colors"
-	"github.com/cupcakearmy/autorestic/internal/lock"
 	"github.com/spf13/cobra"
 )
 
@@ -12,9 +11,9 @@ var checkCmd = &cobra.Command{
 	Short: "Check if everything is setup",
 	Run: func(cmd *cobra.Command, args []string) {
 		internal.GetConfig()
-		err := lock.Lock()
+		err := internal.Lock()
 		CheckErr(err)
-		defer lock.Unlock()
+		defer internal.Unlock()
 
 		CheckErr(internal.CheckConfig())
