+++ cmd/restore.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/cupcakearmy/autorestic/internal"
-	"github.com/cupcakearmy/autorestic/internal/lock"
 	"github.com/spf13/cobra"
 )
 
@@ -14,9 +13,9 @@ var restoreCmd = &cobra.Command{
 	Args:  cobra.MaximumNArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		internal.GetConfig()
-		err := lock.Lock()
+		err := internal.Lock()
 		CheckErr(err)
-		defer lock.Unlock()
+		defer internal.Unlock()
 
 		location, _ := cmd.Flags().GetString("location")
 		l, ok := internal.GetLocation(location)
