+++ cmd/root.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cupcakearmy/autorestic/internal"
 	"github.com/cupcakearmy/autorestic/internal/colors"
 	"github.com/cupcakearmy/autorestic/internal/flags"
-	"github.com/cupcakearmy/autorestic/internal/lock"
 	"github.com/spf13/cobra"
 
 	homedir "github.com/mitchellh/go-homedir"
@@ -18,7 +17,7 @@ import (
 func CheckErr(err error) {
 	if err != nil {
 		colors.Error.Fprintln(os.Stderr, "Error:", err)
-		lock.Unlock()
+		internal.Unlock()
 		os.Exit(1)
 	}
 }
